feat(handle_payments_not_multiple_of_price): Change transcation farmer balances to floats

+++ handler_payments.go
@@ -127,7 +127,7 @@ func (apiCfg *apiConfig) handlerGetPayments(w http.ResponseWriter, r *http.Reque
 		}
 		customPayment.UserName = user.Name
 		customPayment.FarmerName = farmer.Name
-		customPayment.FarmerChickenBalance = farmer.ChickenBalance.Int32
+		customPayment.FarmerChickenBalance = farmer.ChickenBalance.Float64
 		customPayment.FarmerCashBalance = farmer.CashBalance.Int32
 		paymentResponse = append(paymentResponse, customPayment)
 	}
